# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DustNote API
  version: 2.5.12
  description: |
    尘心笔记 REST API。所有笔记内容以 E2EE 密文形式传输，服务端无法解密。
  license:
    name: MIT
servers:
  - url: /api/v1
    description: 同源部署
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NoteRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        ciphertext: { type: string, description: 'E2EE 密文 JSON 字符串' }
        keyVersion: { type: integer }
        isPinned: { type: boolean }
        isFavorite: { type: boolean }
        deletedAt: { type: string, format: 'date-time', nullable: true }
        version: { type: integer, description: '乐观锁版本号' }
        clientUpdatedAt: { type: string, format: 'date-time' }
        serverUpdatedAt: { type: string, format: 'date-time' }
        folderId: { type: string, format: uuid, nullable: true }
      required:
        [
          id,
          ciphertext,
          keyVersion,
          isPinned,
          isFavorite,
          version,
          clientUpdatedAt,
          serverUpdatedAt,
        ]
...
paths:
...
  /notes:
    get:
      summary: 获取笔记列表
      parameters:
        - name: since
          in: query
          schema: { type: string, format: 'date-time' }
          description: 增量同步游标
        - name: includeDeleted
          in: query
          schema: { type: string, enum: ['0', '1'] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items: { $ref: '#/components/schemas/NoteRow' }
    post:
      summary: 创建笔记
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, format: uuid }
                ciphertext: { type: string }
                keyVersion: { type: integer }
                isPinned: { type: boolean }
                isFavorite: { type: boolean }
                clientUpdatedAt: { type: string, format: 'date-time' }
                folderId: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          description: 创建成功
  /notes/{id}:
    patch:
      summary: 更新笔记（乐观锁）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ciphertext: { type: string }
                isPinned: { type: boolean }
                isFavorite: { type: boolean }
                deletedAt: { type: string, nullable: true }
                clientUpdatedAt: { type: string, format: 'date-time' }
                version: { type: integer, description: '当前版本号（乐观锁）' }
                folderId: { type: string, nullable: true }
      responses:
        '200':
          description: 更新成功
        '409':
          description: 版本冲突
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  message: { type: string }
                  current: { $ref: '#/components/schemas/NoteRow' }
    delete:
      summary: 软删除笔记
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 删除成功
...